Add files via upload
This is the beta for my church website. The front end is mostly completed. I will need to add more dynamic areas to it later and eventually U would also like to change some of the design if possible. 

The backend is a complete mess right now, however the general layout is in place.